Move translate to an api service
